requirements.txt
# Core
torch>=2.3.0
transformers>=4.45.0
accelerate>=0.33.0
peft>=0.12.0
bitsandbytes>=0.43.0

# ASR
openai-whisper>=20240930
faster-whisper>=1.0.3
silero-vad>=5.1

# Retrieval / RAG
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0            # swap for faiss-gpu in prod
FlagEmbedding>=1.2.11       # for bge-m3

# TTS
# cosyvoice install from source (see scripts/download_models.py)

# API / streaming
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
pydantic>=2.8.0

# Audio
numpy>=1.26.0
scipy>=1.13.0
soundfile>=0.12.1
librosa>=0.10.2

# Evaluation
sacrebleu>=2.4.0
simuleval>=1.1.4

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.1
loguru>=0.7.2
tqdm>=4.66.0

# Dev
pytest>=8.3.0
pytest-asyncio>=0.23.8
ruff>=0.6.0
black>=24.8.0
